packages/ledger-bridge/requirements.txt
xrpl-py
pika
python-dotenv